requirements.txt
pytest
bandit==1.7.9
black==24.4.2
boto3==1.34.88
botocore==1.34.88
coverage==7.4.4
markdown-it-py==3.0.0
pydantic==2.8.2
moto==5.0.11
mypy-extensions==1.0.0
pg8000
pandas
dlt[postgres]
python-dotenv